flowedit: Extract FlowEdit struct and impl into flow_edit.rs (#2626)
Move FlowEdit struct, message enums, toolbar style functions,
resolve_flow_def_mut, and all impl blocks from main.rs to flow_edit.rs.
main.rs now contains only main(), CLI parsing, and flow loading helpers
(151 lines vs 2957 before). No functional changes — pure code movement.
