+++ flowedit/src/ui_test.rs
@@ -1,10 +1,15 @@
 #![allow(clippy::indexing_slicing, clippy::unwrap_used)]
 
 use super::*;
+use crate::flow_canvas::CanvasMessage;
+use crate::library_panel::{self, LibraryTree};
 use flowcore::model::connection::Connection;
 use flowcore::model::name::HasName;
+use flowcore::model::route::Route;
+use iced::window;
 use iced_test::simulator::{self, simulator};
 use std::collections::HashMap;
+use url::Url;
 
 fn test_win_state() -> WindowState {
     WindowState {
